+++ .atoum.php
@@ -1,6 +1,6 @@
 <?php
 
 /*$runner
-    ->getExtension(mageekguy\atoum\autoloop\extension::class)
+    ->getExtension(atoum\atoum\autoloop\extension::class)
     ->setWatchedFiles(array(__DIR__ . '/classes'))
 ;*/